requirements.txt
# FastAPI dependencies
fastapi
uvicorn

# CORS Middleware for FastAPI
fastapi-cors

# Pydantic for data validation
pydantic

# Python Dotenv for loading environment variables
python-dotenv

# Pydantic Settings for loading environment variables
pydantic-settings

# Motor for MongoDB async client
motor

# Google Cloud AI Platform SDK for using Vertex AI models
google-cloud-aiplatform

# Vertex AI SDK for accessing specific models
vertexai

# LangChain for working with document loaders, retrievers, and LLMs
vertexai

# Google Generative AI Embeddings for vector generation
langchain_google_genai

# LangChain Community for additional functionality
langchain_community
